adjust kinopoisk Upcoming / PopularMovies fetching

+++ MyMovies/Network/ResponseTypes/kinopoisk/KinopoiskMovieResponseProtocol.swift
@@ -28,7 +28,7 @@ protocol KinopoiskGenreResponseProtocol: Codable {
 }
 
 protocol KinopoiskRatingResponseProtocol: Codable {
-    var kp: Float? { get }
+    var kp: Double? { get }
 }
 
 protocol KinopoiskCoverResponseProtocol: Codable {
